Break apart Orchestration handlers.py
