+++ containers/orchestration/app/handlers/tracer.py
@@ -1,0 +1,4 @@
+from opentelemetry import trace
+
+# Integrate services tracer with automatic instrumentation context
+tracer = trace.get_tracer("orchestration_handlers.py_tracer")
