+++ containers/orchestration/tests/test_handlers.py
@@ -4,21 +4,29 @@ from pathlib import Path
 from unittest.mock import MagicMock
 from unittest.mock import Mock
 
-from app.handlers import build_fhir_converter_request
-from app.handlers import build_geocoding_request
-from app.handlers import build_ingestion_dob_request
-from app.handlers import build_ingestion_name_request
-from app.handlers import build_ingestion_phone_request
-from app.handlers import build_message_parser_message_request
-from app.handlers import build_message_parser_phdc_request
-from app.handlers import build_stamp_condition_extensions_request
-from app.handlers import build_validation_request
-from app.handlers import unpack_fhir_converter_response
-from app.handlers import unpack_fhir_to_phdc_response
-from app.handlers import unpack_ingestion_standardization
-from app.handlers import unpack_parsed_message_response
-from app.handlers import unpack_stamp_condition_extensions_response
-from app.handlers import unpack_validation_response
+from app.handlers.request_builders.fhir_converter import build_fhir_converter_request
+from app.handlers.request_builders.ingestion import build_geocoding_request
+from app.handlers.request_builders.ingestion import build_ingestion_dob_request
+from app.handlers.request_builders.ingestion import build_ingestion_name_request
+from app.handlers.request_builders.ingestion import build_ingestion_phone_request
+from app.handlers.request_builders.ingestion import build_validation_request
+from app.handlers.request_builders.message_parser import (
+    build_message_parser_message_request,
+)
+from app.handlers.request_builders.message_parser import (
+    build_message_parser_phdc_request,
+)
+from app.handlers.request_builders.trigger_code_reference import (
+    build_stamp_condition_extensions_request,
+)
+from app.handlers.response_builders.fhir_converter import unpack_fhir_converter_response
+from app.handlers.response_builders.ingestion import unpack_ingestion_standardization
+from app.handlers.response_builders.ingestion import unpack_validation_response
+from app.handlers.response_builders.message_parser import unpack_fhir_to_phdc_response
+from app.handlers.response_builders.message_parser import unpack_parsed_message_response
+from app.handlers.response_builders.trigger_code_reference import (
+    unpack_stamp_condition_extensions_response,
+)
 from lxml import etree
